refactor: Merge testingTB to mockingT to reduce code
