+++ test_test.go
@@ -4,33 +4,22 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"runtime"
 	"strings"
+	"sync"
 	"testing"
 	"time"
 
 	"github.com/raeperd/test"
 )
 
-type testingTB struct {
-	testing.TB
-	failed bool
-	w      io.Writer
-}
-
-func (t *testingTB) Helper() {}
-
-func (t *testingTB) Fatalf(format string, args ...any) {
-	t.failed = true
-	fmt.Fprintf(t.w, format, args...)
-}
-
 func Test(t *testing.T) {
 	beOkay := func(callback func(tb testing.TB)) {
 		t.Helper()
 		var buf strings.Builder
-		tb := &testingTB{w: &buf}
-		callback(tb)
-		if tb.failed {
+		mt := &mockingT{w: &buf}
+		callback(mt)
+		if mt.Failed() {
 			t.Fatal("failed too soon")
 		}
 		if buf.String() != "" {
@@ -47,9 +36,9 @@ func Test(t *testing.T) {
 	beBad := func(callback func(tb testing.TB)) {
 		t.Helper()
 		var buf strings.Builder
-		tb := &testingTB{w: &buf}
-		callback(tb)
-		if !tb.failed {
+		mt := &mockingT{w: &buf}
+		callback(mt)
+		if !mt.Failed() {
 			t.Fatal("did not fail")
 		}
 		if buf.String() == "" {
@@ -64,4 +53,67 @@ func Test(t *testing.T) {
 	beBad(func(tb testing.TB) { test.NotNil(tb, nil) })
 	beBad(func(tb testing.TB) { test.True(tb, false) })
 	beBad(func(tb testing.TB) { test.False(tb, true) })
+}
+
+type mockingT struct {
+	testing.T
+	sync.RWMutex
+	failed   bool
+	cleanups []func()
+	w        io.Writer
+}
+
+func (m *mockingT) setFailed(b bool) {
+	m.Lock()
+	defer m.Unlock()
+	m.failed = b
+}
+
+func (m *mockingT) Failed() bool {
+	m.RLock()
+	defer m.RUnlock()
+	return m.failed
+}
+
+func (m *mockingT) Run(name string, f func(t *testing.T)) {
+	m.cleanups = nil
+	m.setFailed(false)
+	ch := make(chan struct{})
+	defer func() {
+		for _, f := range m.cleanups {
+			defer f()
+		}
+	}()
+	// Use a goroutine so Fatalf can call Goexit
+	go func() {
+		defer close(ch)
+		f(&m.T)
+	}()
+	<-ch
+}
+
+func (m *mockingT) Cleanup(f func()) {
+	m.cleanups = append(m.cleanups, f)
+}
+
+func (*mockingT) Log(args ...any) {
+	fmt.Println(args...)
+}
+
+func (*mockingT) Helper() {}
+
+func (m *mockingT) Fatalf(format string, args ...any) {
+	m.setFailed(true)
+	if m.w != nil {
+		fmt.Fprintf(m.w, format, args...)
+		// Do not call runtime.Goexit here, so that caller can read the output
+	} else {
+		m.Errorf(format, args...)
+		runtime.Goexit()
+	}
+}
+
+func (m *mockingT) Errorf(format string, args ...any) {
+	m.setFailed(true)
+	fmt.Printf(format+"\n", args...)
 }
